requirements-gpu.txt
pandas
numpy
scipy
matplotlib
scikit-learn
statsmodels
tensorflow-gpu=1.09.0
h5py
seaborn
